move FromNaslValue traits impl to packet forgery
