+++ rust/src/nasl/builtin/raw_ip/packet_forgery.rs
@@ -38,7 +38,6 @@ use pnet::packet::{
     ipv6::{Ipv6Packet, MutableIpv6Packet},
     tcp::{TcpOption, TcpOptionNumbers, TcpPacket, *},
     udp::{MutableUdpPacket, UdpPacket},
-    Packet, PrimitiveValues,
 };
 
 use pnet_macros_support::types::u9be;
@@ -118,6 +117,112 @@ impl<'a> FromNaslValue<'a> for PacketPayload {
             NaslValue::String(d) => Ok(d.as_bytes().to_vec().into()),
             NaslValue::Number(d) => Ok(d.to_be_bytes().to_vec().into()),
             _ => Ok(Vec::<u8>::new().into()),
+        }
+    }
+}
+
+impl<'a> FromNaslValue<'a> for Ipv4Packet<'a> {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let buf: &[u8] = <&[u8]>::from_nasl_value(val)?;
+        let ip = Ipv4Packet::new(buf).ok_or(BuiltinError::RawIp(RawIpError::PacketForgery(
+            PacketForgeryError::CreateIpv4Packet,
+        )))?;
+        Ok(ip)
+    }
+}
+
+impl<'a> FromNaslValue<'a> for Ipv6Packet<'a> {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let buf: &[u8] = <&[u8]>::from_nasl_value(val)?;
+        let ip = Ipv6Packet::new(buf).ok_or(BuiltinError::RawIp(RawIpError::PacketForgery(
+            PacketForgeryError::CreateIpv6Packet,
+        )))?;
+        Ok(ip)
+    }
+}
+
+fn get_payload_from_packet(val: &NaslValue) -> Result<Vec<u8>, FnError> {
+    // It expects an Ipv4 or Ipv6 packet. Before to get a packet to the right
+    // version, we need to know the version. So, we extract this from the header.
+    let first_4bits = (<&[u8]>::from_nasl_value(val)?[0] & 0b11110000) >> 4;
+    match first_4bits {
+        4 => {
+            let ip: Ipv4Packet = Ipv4Packet::from_nasl_value(val)?;
+            Ok(ip.payload().to_vec())
+        }
+        6 => {
+            let ip: Ipv6Packet = Ipv6Packet::from_nasl_value(val)?;
+            Ok(ip.payload().to_vec())
+        }
+        _ => Err(BuiltinError::RawIp(RawIpError::PacketForgery(
+            PacketForgeryError::CreateTcpPacket,
+        ))
+        .into()),
+    }
+}
+
+impl<'a> FromNaslValue<'a> for UdpPacket<'a> {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let payload = get_payload_from_packet(val)?;
+        let udp = UdpPacket::owned(payload).ok_or(BuiltinError::RawIp(
+            RawIpError::PacketForgery(PacketForgeryError::CreateUdpPacket),
+        ))?;
+        Ok(udp)
+    }
+}
+
+impl<'a> FromNaslValue<'a> for TcpPacket<'a> {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let payload = get_payload_from_packet(val)?;
+        let tcp = TcpPacket::owned(payload).ok_or(BuiltinError::RawIp(
+            RawIpError::PacketForgery(PacketForgeryError::CreateTcpPacket),
+        ))?;
+        Ok(tcp)
+    }
+}
+
+impl<'a> FromNaslValue<'a> for IcmpPacket<'a> {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let payload = get_payload_from_packet(val)?;
+        let icmp = IcmpPacket::owned(payload).ok_or(BuiltinError::RawIp(
+            RawIpError::PacketForgery(PacketForgeryError::CreateIcmpPacket),
+        ))?;
+        Ok(icmp)
+    }
+}
+
+impl<'a> FromNaslValue<'a> for Icmpv6Packet<'a> {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let payload = get_payload_from_packet(val)?;
+        let icmp = Icmpv6Packet::owned(payload).ok_or(BuiltinError::RawIp(
+            RawIpError::PacketForgery(PacketForgeryError::CreateIcmpPacket),
+        ))?;
+        Ok(icmp)
+    }
+}
+
+impl<'a> FromNaslValue<'a> for Ipv4Addr {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let addr = String::from_nasl_value(val)?;
+        match addr.parse::<Ipv4Addr>() {
+            Ok(ip_addr) => Ok(ip_addr),
+            Err(e) => Err(ArgumentError::WrongArgument(
+                format!("Expected a valid IPv4 address. {}.", e).to_string(),
+            )
+            .into()),
+        }
+    }
+}
+
+impl<'a> FromNaslValue<'a> for Ipv6Addr {
+    fn from_nasl_value(val: &'a NaslValue) -> Result<Self, FnError> {
+        let addr = String::from_nasl_value(val)?;
+        match addr.parse::<Ipv6Addr>() {
+            Ok(ip_addr) => Ok(ip_addr),
+            Err(e) => Err(ArgumentError::WrongArgument(
+                format!("Expected a valid IPv6 address. {}.", e).to_string(),
+            )
+            .into()),
         }
     }
 }
